refactor: move all stream creation form logic to own component
